+++ eth_defi/erc_4626/classification.py
@@ -35,11 +35,16 @@ class VaultFeatureProbe:
 def create_probe_calls(
     addresses: Iterable[HexAddress],
     share_probe_amount=1_000_000,
+    chain_id: int | None = None,
 ) -> Iterable[EncodedCall]:
     """Create calls that call each vault address using multicall.
 
     - Because ERC standards are such a shit show, and nobody is using good interface standard,
       we figure out the vault type by probing it with various calls
+
+    :param chain_id:
+        Limit probes by a chain, so that we do not try to probe vaults that exist only on certain
+        chains like mainnet.
     """
 
     convert_to_shares_payload = eth_abi.encode(["uint256"], [share_probe_amount])
@@ -692,6 +697,12 @@ def detect_vault_features(
 
     assert address.lower() not in BROKEN_VAULT_CONTRACTS, f"Vault {address} is known broken vault contract like, avoid"
 
+    hardcoded_flags = HARDCODED_PROTOCOLS.get(address.lower())
+    if hardcoded_flags:
+        features = hardcoded_flags
+        logger.debug("Using hardcoded vault features for %s: %s", address, features)
+        return hardcoded_flags
+
     address = Web3.to_checksum_address(address)
     logger.info("Detecting vault features for %s", address)
     probe_calls = list(create_probe_calls([address]))
@@ -805,6 +816,11 @@ def create_vault_instance(
         from eth_defi.untangle.vault import UntangleVault
 
         return UntangleVault(web3, spec, token_cache=token_cache, features=features)
+    elif ERC4626Feature.cap_like in features:
+        # Covered Agent Protocol (CAP) uses Yearn V3 infrastructure
+        from eth_defi.cap.vault import CAPVault
+
+        return CAPVault(web3, spec, token_cache=token_cache, features=features)
     elif ERC4626Feature.yearn_v3_like in features or ERC4626Feature.yearn_tokenised_strategy in features:
         # Both of these have fees internatilised
         from eth_defi.yearn.vault import YearnV3Vault
@@ -881,3 +897,13 @@ def create_vault_instance_autodetect(
     vault = create_vault_instance(web3, vault_address, features=features, token_cache=token_cache)
     assert vault is not None, f"Could not create vault instance: {vault_address} with features {features}"
     return vault
+
+
+#: Handle problematic protocols.
+#:
+#: Some protocols cannot be detected by their vault smart contract structure, because they are using copy-paste smart contracts.
+#: For these, we need to do by vault contract address whitelisting here.
+#:
+HARDCODED_PROTOCOLS = {
+    "0x3ed6aa32c930253fc990de58ff882b9186cd0072": {ERC4626Feature.cap_like},
+}