+++ eth_defi/erc_4626/vault.py
@@ -1,22 +1,19 @@
 """Generic ECR-4626 vault reader implementation."""
 
 import datetime
-
 import logging
-
 from decimal import Decimal
 from functools import cached_property
-from typing import Iterable, TypeAlias, Literal
+from typing import Iterable, Literal, TypeAlias
 
 import eth_abi
 from eth_typing import HexAddress
+from requests.exceptions import HTTPError
 from web3 import Web3
 from web3.contract import Contract
+
 from eth_defi.compat import WEB3_PY_V7
 from eth_defi.provider.fallback import ExtraValueError
-
-from requests.exceptions import HTTPError
-
 from eth_defi.vault.flag import VaultFlag
 
 if WEB3_PY_V7:
@@ -28,11 +25,11 @@ from web3.types import BlockIdentifier
 
 from eth_defi.abi import ZERO_ADDRESS_STR
 from eth_defi.balances import fetch_erc20_balances_fallback, fetch_erc20_balances_multicall
-from eth_defi.erc_4626.core import get_deployed_erc_4626_contract, ERC4626Feature
-from eth_defi.event_reader.conversion import convert_int256_bytes_to_int, convert_uint256_bytes_to_address, BadAddressError
-from eth_defi.event_reader.multicall_batcher import EncodedCall, EncodedCallResult, BatchCallState
+from eth_defi.erc_4626.core import ERC4626Feature, get_deployed_erc_4626_contract
+from eth_defi.event_reader.conversion import BadAddressError, convert_int256_bytes_to_int, convert_uint256_bytes_to_address
+from eth_defi.event_reader.multicall_batcher import BatchCallState, EncodedCall, EncodedCallResult
 from eth_defi.token import TokenDetails, fetch_erc20_details, is_stablecoin_like
-from eth_defi.vault.base import VaultBase, VaultSpec, VaultInfo, TradingUniverse, VaultPortfolio, VaultFlowManager, VaultHistoricalReader, VaultHistoricalRead
+from eth_defi.vault.base import TradingUniverse, VaultBase, VaultFlowManager, VaultHistoricalRead, VaultHistoricalReader, VaultInfo, VaultPortfolio, VaultSpec
 
 logger = logging.getLogger(__name__)